Add one-click plugin update, update checker, test suite, and bug fixes
- Built-in update system: check GitHub releases, download DLL, atomic
  file replacement with backup/rollback, restart Emby — all from the
  Dashboard tab
- Persist LastInstalledVersion in config to suppress update banner after
  restart
- Docker compatibility: fall back to ApplicationPaths.PluginsPath when
  Assembly.Location is empty
- Automatic git-tag versioning in build.sh (e.g. v1.1.0 → 1.1.0,
  v1.1.0-3-gabcdef → 1.1.0.3)
- Extract log sanitizer into reusable LogSanitizer service
- Fix Dispatcharr bitrate parsing (int → double)
- Add test project with 115 tests covering UpdateChecker, LogSanitizer,
  DispatcharrClient, ChannelNameCleaner, ContentNameCleaner, and
  FolderMappingParser

+++ Emby.Xtream.Plugin/Client/Models/StreamStatsInfo.cs
@@ -17,6 +17,6 @@ namespace Emby.Xtream.Plugin.Client.Models
         public double? SourceFps { get; set; }
 
         [JsonPropertyName("ffmpeg_output_bitrate")]
-        public int? Bitrate { get; set; }
+        public double? Bitrate { get; set; }
     }
 }
